[System]: Make HttpListener partial and move Mono-specific bits into HttpListener.Mono.cs. (#4808)
* [System]: Make HttpListener partial and move Mono-specific bits into HttpListener.Mono.cs.

* Remove ancient EMBEDDED_IN_1_0 conditional.

* Remove unused using statements.

* HttpListener.CreateSslStream() now uses RemoteCertificateValidationCallback.
