+++ mcs/class/System/System.Net/HttpListenerPrefixCollection.cs
@@ -25,19 +25,13 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
-#if SECURITY_DEP || EMBEDDED_IN_1_0
+#if SECURITY_DEP
 
 using System.Collections;
 using System.Collections.Generic;
 namespace System.Net {
-#if EMBEDDED_IN_1_0
-	public class HttpListenerPrefixCollection : IEnumerable, ICollection {
-		ArrayList prefixes;
-		
-#else
 	public class HttpListenerPrefixCollection : ICollection<string>, IEnumerable<string>, IEnumerable {
 		List<string> prefixes = new List<string> ();
-#endif
 		HttpListener listener;
 
 		internal HttpListenerPrefixCollection (HttpListener listener)
@@ -95,12 +89,10 @@ namespace System.Net {
 			((ICollection) prefixes).CopyTo (array, offset);
 		}
 
-#if !EMBEDDED_IN_1_0
 		public IEnumerator<string> GetEnumerator ()
 		{
 			return prefixes.GetEnumerator ();
 		}
-#endif
 	
 		IEnumerator IEnumerable.GetEnumerator ()
 		{
