+++ mcs/class/System/common.sources
@@ -62,6 +62,7 @@ System.Net/FtpWebRequest.cs
 System.Net/FtpWebResponse.cs
 System.Net/HttpConnection.cs
 System.Net/HttpListener.cs
+System.Net/HttpListener.Mono.cs
 System.Net/HttpListenerBasicIdentity.cs
 System.Net/HttpListenerContext.cs
 System.Net/HttpListenerPrefixCollection.cs
