+++ mcs/class/System/monotouch_watch_System.dll.exclude.sources
@@ -32,6 +32,7 @@ System.Net/FtpWebRequest.cs
 System.Net/FtpWebResponse.cs
 System.Net/HttpConnection.cs
 System.Net/HttpListener.cs
+System.Net/HttpListener.Mono.cs
 System.Net/HttpListenerContext.cs
 System.Net/HttpListenerPrefixCollection.cs
 System.Net/HttpListenerRequest.cs
